Fix: Domain 으로 Package 구조 변경
