+++ src/main/java/com/creativesemester/SejongCodingMate/domain/member/repository/MemberRepository.java
@@ -1,7 +1,7 @@
-package com.creativesemester.SejongCodingMate.member.repository;
+package com.creativesemester.SejongCodingMate.domain.member.repository;
 
 
-import com.creativesemester.SejongCodingMate.member.entity.Member;
+import com.creativesemester.SejongCodingMate.domain.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
