+++ src/main/java/com/creativesemester/SejongCodingMate/global/security/UserDetailsImpl.java
@@ -1,6 +1,6 @@
 package com.creativesemester.SejongCodingMate.global.security;
 
-import com.creativesemester.SejongCodingMate.member.entity.Member;
+import com.creativesemester.SejongCodingMate.domain.member.entity.Member;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
